Self-Assembler Written
Self assembler is finally done, but it needs to be bug tested. It currently loops right after the first 3 comments of TestProg. Would recommend using a bunch of breakpoints via "HLT; 0x0[Breakpoint number]." It also might help to have a copy of the symbol table, and have compressed copies of files on hand so that PC numbers could be directly matched to source code lines.