Important CPU I/O Fix
Fixed subtle issue with CPU Read/Write I/O pins going high when they shouldn't. This is a fix for the original cause of #9.

Changed RTL tab characters to '\t' as opposed to '    '

Note: I noticed that the self-modifying code portion of the assembler did not work properly. This is probably a CPU issue as well.