Working Self Assembler
Resolved #9. The assembler is working!!! ...but it's too slow. Specifically GetNext calling TblIdxLookup all the time is very slow.

The simple assembler test file was able to be assembled in 72148 clock cycles. I would like this number at least halved before I'm willing to call it acceptable.

Changelog:

Fixed issue in HashBuff (it selected the 5th bit instead of the 6th one).

Fixed missing argument error in GetNext (it didn't set RG2 to 1 before calling TblIdxLookup).

Moved comments around in the self-assembler to be more consistant.

Label table is now written to a file instead of printed in TestingAssembler.py

RTL and table file suffixes are now constants in TestingAssembler.py

Fixed significant problem in BuildTables in the self-assembler (for some reason I wrote MOV RG1 IO as opposed to calling GetNext).

Fixed issue in LabelLookup where BuffIdx was not loaded, but was assumed to be in RG3.