Migrate from Logisim to Digital
Very Important: The Logisim version of the CPU is no longer being worked on. Some of these changes have not been applied to it. Some of these changes are rather important, too!

But wow! 50KHz!! I never thought I'd see the assembler run that quickly in a simulation!

Resolved #16 !!

And fittingly, it took the self-assembler about 16 seconds to assemble itself... uncompressed!

Changelog:

Bound carry out of ENC to any bits being on. To put it in TASL, these now mean the same thing:
```
PAS RG0; JIF EQZ >RG0Zero

ALU ENC RG0 RG0; JIF NCR >RG0Zero
```

Moved I/O R/W signal corrections to the CPU, as the Registers & Selection circuit should not be in charge of _how_ the CPU communicates with I/O, only _that_ it communicates.

CTD now has to be turned off (ExEn = 0) to be written to if equal to 0. Not sure how you could have written to it anyways if this was the case, since this would cause a CTDZ exception.

Clear EXC Buffer in the Flags circuit is now synchronous as opposed to asynchronous. However, it appears that it was already synchronous to a certain extent.

Split the control unit into two parts: The CISC Decoder, and the RISC Controller. (also, and this is news to me, I made a CISC CPU???)

Fixed a critical problem where, if an instruction was on state 1/2 and you Paused the CPU, the Fetch register would be updated with the current RAM output as an instruction.

Fetch register is now bypassed on state 0 using a multiplexer.

Exception buffer now updates on falling edge.

Fixed the overflow of NEG to be (Sign(OUT) XNOR Sign(B)) instead of (Sign(OUT) XNOR Sign(A))

Hard Reset no longer resets everything, it only resets some flags, PC, and the state counter.

Added rising edge monostable circuits to the start and pause inputs inside the CPU.

Fixed issue where NOP causes PRIV exception.

Changed scripts to output .hex files when a file should be loaded into simulation memory (Digital does bizarre things if the file is not named .hex.

Removed New Mnemonics Test

Added Sort Test

Numbered some of the folders

Removed tab replacement in the FTLV1 (since Digital's terminal supports tab characters).

Finished a comprehensive piece of documentation on the CPU in Overleaf.

Moved some files into old, including the old Logisim hardware.