Bootloader Done
Resolved #15: Bootloader is finished and tested and integrated into the assembler computer!

Added new file format to BOTH assemblers which starts the file with the number of words in the file. For Python it's an argument "-c". For TASL its the constant boolean GENLOADABLE.

Resolved #18: Added string support to TASL
```
#MyString = "Hello World"
```

Added INC increment and DEC decrement mnemonics (that also set the flags)

Updated TASL xml file and btn file

Successfully ran Hello World for the first time

Fixed CPU issue which internally swapped HI and LO