New Test, Assembler Tweaks, More OS Pseudocode
Telescope Computer Test created

Expanded the list of important characters to &, $, and !

Created better (and accurate) table of important characters

Fixed issue where ? was treated as important in the python assembler, but not in the self-assembler (it was too difficult to add to the self-assembler, so ? is no longer important).

More work done on TBOS 1.0 pseudocode. It's mostly done.