Test Written, Computer Mostly Working, New TASL Feature
I finished writing the Telescope Computer test. The computer passes it almost entirely, except there seems to be an issue with the TNET chips. I'm not exactly sure what it is though. The computer doesn't seem to be able to read or write anything from them.

I fixed some miscellaneous hardware issues with the computer (pausing not working as expected, port bits were off from their devices, etc.)

TASL has a new feature, which lets you use labels in an immediate ALU operation (though it will be masked to 4 bits). This also lets you define constants which are the value of other constants before it in the code.